exists追加

+++ backend/transactions/rooms/delete_rooms_transaction.py
@@ -1,3 +1,8 @@
+from fastapi import HTTPException
+
+from exists.exists_devices_in_rooms import (
+    exists_devices_in_rooms
+)
 from rooms.delete_rooms import delete_rooms
 from schemas.room_schemas import DeleteRoomsRequest
 
@@ -7,6 +12,16 @@ def delete_room_transaction(
                             ):
 
     print("delete room transaction")
+    if exists_devices_in_rooms(
+        room.ids,
+        hospital_id
+    ):
+      raise HTTPException(
+        status_code=400,
+        detail="機器が配置されている部屋は削除できません。"
+      )
+
+
 
     delete_rooms(
                   room=room,
